fix: revert previous commit, use lead data
v1.0.0 would dynamically generate the bundle while the game was running. I
cannot find a way to time this generation where:

- a locale has been loaded
- the game's content has not been loaded

This caused the data scraper by Pulverizer to not include the needed keys.
Instead, I'm going to revert back to how I did it previously where it was
generated externally. Note that I've rewritten it to use lead's data and
therefore it now supports every single Mindustry version.